+++ gatsby-browser.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import Layout from './src/components/general/Layout'
+import Layout from './src/components/Layout'
 
 const transitionDelay = 600
