+++ gatsby-ssr.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import Layout from './src/components/general/Layout'
+import Layout from './src/components/Layout'
 
 export const wrapPageElement = ({ element, props }) => {
   return <Layout {...props}>{element}</Layout>
