+++ src/components/Footer.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import { Link } from 'gatsby'
-import { Wrapper, Content } from '..//general/Utils'
+import { Wrapper, Content } from './Utils'
 import { FiTwitter, FiInstagram } from 'react-icons/fi'
 
 const Footer = props => {
