+++ src/components/galleryGrid.js
